RDKB-11810,RDKB-11811 : RxBcastPowerLevel and RxSNR inconsistent between nodes
Reason for change: RxBcastPowerLevel type changed to int and memset structure to 0.
Test Procedure:
-> Check Device.MoCA.Interface.1.AssociatedDevice.1.RxBcastPowerLevel
It should be "int". Also, it can show negative value. It should have consistent values among the nodes.
-> Check Device.MoCA.Interface.1.AssociatedDevice.1.RxSNR
It should have consistent values among the nodes.
Risks: None

+++ source/TR-181/include/cosa_moca_apis.h
@@ -358,7 +358,7 @@ struct _COSA_DML_MOCA_ASSOC_DEVICE
     ULONG                           TxPowerControlReduction;
     INT                             RxPowerLevel;
     ULONG                           TxBcastRate;
-    ULONG                           RxBcastPowerLevel;
+    INT                             RxBcastPowerLevel;
     ULONG                           TxPackets;
     ULONG                           RxPackets;
     ULONG                           RxErroredAndMissedPackets;
